MultiIndexPopulator force flush all populators after store scan is complete
Needed because MultiIndexPopulator batches populator updates when scanning store
and last batch may not have been applied at end of store scan.
